refactor: separate serializers
