+++ app/serializers/__init__.py
@@ -1,0 +1,5 @@
+from app.serializers.course import *
+from app.serializers.similarity import *
+from app.serializers.submission import *
+from app.serializers.task import *
+from app.serializers.user import *
