+++ app/serializers/course.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from app.models import Course
+
+
+class CourseSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Course
+        fields = "__all__"
